+++ packages/unmock-node/src/__tests__/backend/states.test.ts
@@ -1,7 +1,8 @@
 import axios from "axios";
 import path from "path";
-import { CorePackage } from "unmock-core";
-import { dsl } from "../..";
+import { CorePackage, States } from "unmock-core";
+import { dsl, Request } from "../../..";
+
 import NodeBackend from "../../backend";
 
 class StateTestPackage extends CorePackage {
@@ -16,7 +17,7 @@ describe("Node.js interceptor", () => {
   describe("with state requests in place", () => {
     let nodeInterceptor: NodeBackend;
     let unmock: StateTestPackage;
-    let states: any;
+    let states: States;
 
     beforeAll(() => {
       nodeInterceptor = new NodeBackend({ servicesDirectory });
@@ -26,7 +27,6 @@ describe("Node.js interceptor", () => {
 
     afterAll(() => {
       unmock.off();
-      states = undefined;
     });
 
     beforeEach(() => states.reset());
@@ -106,6 +106,12 @@ describe("Node.js interceptor", () => {
       states.petstore(() => "baz");
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.data).toBe("baz");
+    });
+
+    test("sets an entire response from with request object", async () => {
+      states.petstore((req: Request) => req.host);
+      const response = await axios("http://petstore.swagger.io/v1/pets");
+      expect(response.data).toBe("petstore.swagger.io");
     });
 
     test("sets an entire response from function with DSL", async () => {