+++ packages/unmock-node/src/types.ts
@@ -1,0 +1,5 @@
+// Types and aliases exported from unmock-node
+export {
+    ISerializedRequest as Request,
+    States,
+} from "unmock-core";